adding vec and tok

+++ code/cofi/run.py
@@ -132,10 +132,10 @@ def main():
             revision=model_args.model_revision,
             use_auth_token=True if model_args.use_auth_token else None,
         )
-        config.save_pretrained(os.path.join(training_args.output_dir, 'config.json'))
-        print(f"Saved config to ", os.path.join(training_args.output_dir, 'config.json'))
-        return
-
+        if not os.path.exists(os.path.join(training_args.output_dir, 'config.json')):
+            config.save_pretrained(os.path.join(training_args.output_dir, 'config.json'))
+            print(f"Saved config to ", os.path.join(training_args.output_dir, 'config.json'))
+        
         tokenizer = AutoTokenizer.from_pretrained(
             model_args.model_name_or_path,
             cache_dir=model_args.cache_dir,
